// freeboard_frontend/src/components/units/boards/detail/BoardDetail.types.ts
import { IQuery } from "../../../../commons/types/generated/types";

export interface IBoardDetailUIProps {
  data?: Pick<IQuery, "fetchBoard">;
  onClickMoveToBoardList: () => void;
  onClickMoveToBoardEdit: () => void;
  onClickDeleteBoard: () => void;
  onClickLike: () => void;
  onClickDislike: () => void;
}